requirements.txt
Flask==2.0.3
Flask-Cors==3.0.10
psycopg2-binary==2.9.3
python-dotenv==1.0.0
Flask-Caching==2.0.0
requests==2.31.0
gunicorn==21.2.0
